requirements.txt
PyYAML
pyserial